Rename to sasslint and add config support
The `sass_lint` plugin was added upstread in
scrooloose/syntastic@551bcbcc30dc14e5d6176b6272cc69f70611ef0a
so to avoid the naming conflict the checker has been renamed to
`sasslint`

This commit also adds support sass-lint configs. By default, it will
look for .sass-lint.yml in a parent directory or a path to a specific
config can be set as either a global or a buffer variable.